prototypes/ai-agent-control-center/youtube-processor/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
youtube-transcript-api==0.6.2
pydantic==2.10.3
python-dotenv==1.0.1
